Below for description of bug (found on pthread3)
Hi Kapil,
    And I was thinking about this some more.

First of all, I'm only seeing the bug with popen1 on Discovery (CentOS 7) -- not on Ubuntu/WSL.

As I look at the data, I believe that what happens somehow is:
  1.  dmtcp_command -c   : starts a checkpoint
  2.  Once the coordinator state reaches CHECKPOINTED, then somehow a signal is sent to the user thread of popen1.  So, the user thread wakes up from its sleep.  And so, the user thread has a DMTCP_EVENT_FORK_PREPARE event before calling fork(), and so it sends a message to the coordinator.
  3.  The coordinator then somehow gets confused between a reply message from the checkpoint thread and a message relating to DMTCP_EVENT_FORK_PREPARE, sent by the user thread.

Maybe this spurious signal is sent only on CentOS 7 (on Linux 3.10).

===
And while I was working on this, I found a weird bug with pthread3, pthread4, pthread5.  In fact, this bug is actually exhibited on the risc-v machine.
To see the bug, use 'git apply' to apply the git patch 'bug-pthread.patch' to DMTCP.
It adds a 'sleep(5)' to waitForBarrier() if the barrier is  "DMT:WriteCkpt".  Normally, a sleep() statement should not create a bug.  But the pthread3, pthread4 and phread5 tests fail during checkpointing.

So, this is another bug that I don't have time to diagnose now.  :-)  But it's probably related to the coordinator.

Best,